Recovery does not work at all, not sure why
I've tried a simple DPA so that it can be easily shown. NOTHING works. I
get complete garbage output and have no idea why. Maybe I should do it
correctly and turn to CPA, but DPA is so much easier to explain. Not
sure if I can get it working.